feat(vmx-106): rkyv binary IPC channel for cef<->bevy bridge
- Vendor bevy_cef_core 0.5.2 fork with new BinEmit handler + emit_event_bytes
- Add cef.binEmit V8 native + bin-host-emit/bin-js-emit dispatch
- New BinHostEmitEvent + BinJsEmitEventPlugin in bevy_cef patch
- Wasm helpers: try_cef_bin_emit_rkyv, decode_bin_host_emit_js, use_bin_event_listener
- Migrate all bridge events: theme, command bar, sessions, layout, terminal, processes, UiReady, history
- Allow zero-length payloads for unit-shaped events (UiReady)

+++ crates/vmux_command/src/results.rs
@@ -46,7 +46,7 @@ fn session_result(session: &CommandBarSession) -> CommandBarResultItem {
         name: session.name.clone(),
         profile: session.profile.clone(),
         is_active: session.is_active,
-        tab_count: session.tab_count,
+        tab_count: session.tab_count as usize,
     }
 }
 
@@ -129,7 +129,7 @@ pub fn filter_results(
             title: t.title.clone(),
             url: t.url.clone(),
             pane_id: t.pane_id,
-            tab_index: t.tab_index,
+            tab_index: t.tab_index as usize,
         }));
         items.extend(command_results(commands));
         return items;
@@ -184,7 +184,7 @@ pub fn filter_results(
                     title: t.title.clone(),
                     url: t.url.clone(),
                     pane_id: t.pane_id,
-                    tab_index: t.tab_index,
+                    tab_index: t.tab_index as usize,
                 });
             }
         }
